Moving macros tests into api package
