+++ net.dougqh.jak.jvm.assembler/test-src/net/dougqh/jak/jvm/assembler/JakAssemblerTestSuite.java
@@ -1,7 +1,7 @@
 package net.dougqh.jak.jvm.assembler;
 
 import net.dougqh.jak.jvm.assembler.api.JakAssemblerApiTestSuite;
-import net.dougqh.jak.jvm.assembler.macros.MacrosTestSuite;
+import net.dougqh.jak.jvm.assembler.macros.api.MacrosTestSuite;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
