+++ net.dougqh.jak.jvm.assembler/test-src/net/dougqh/jak/jvm/assembler/api/JakAssemblerApiTestSuite.java
@@ -1,6 +1,6 @@
 package net.dougqh.jak.jvm.assembler.api;
 
-import net.dougqh.jak.jvm.assembler.macros.MacrosTestSuite;
+import net.dougqh.jak.jvm.assembler.macros.api.MacrosTestSuite;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
