+++ net.dougqh.jak.jvm.assembler/test-src/net/dougqh/jak/jvm/assembler/macros/api/IfTest.java
@@ -1,4 +1,4 @@
-package net.dougqh.jak.jvm.assembler.macros;
+package net.dougqh.jak.jvm.assembler.macros.api;
 
 import static net.dougqh.jak.Jak.*;
 import static net.dougqh.jak.assembler.JakAsm.*;
@@ -6,6 +6,7 @@ import static org.hamcrest.CoreMatchers.*;
 import static org.junit.Assert.*;
 import net.dougqh.jak.jvm.assembler.JvmClassWriter;
 import net.dougqh.jak.jvm.assembler.JvmWriter;
+import net.dougqh.jak.jvm.assembler.macros.stmt;
 
 import org.junit.Test;
